# Optimize In by specializing for concrete slice types