Most uses of In (or similar functions) use []int, []interface{} or []string in my experience.

Specialize the append logic for these 3 types. For []interface{} some allocations are avoided and. For []int and []string the code still needs most allocations, but I already have another idea to remove most allocations, which could be implemented in another PR.

Also adds 3 new benchmarks with big slices of these types.

Benchstat:

```
name          old time/op    new time/op    delta
In-8             733ns ± 0%     700ns ± 1%   -4.57%  (p=0.000 n=9+10)
In1k-8          36.1µs ± 0%    17.4µs ± 1%  -51.69%  (p=0.000 n=10+10)
In1kInt-8       58.5µs ± 0%    40.4µs ± 0%  -30.92%  (p=0.000 n=10+9)
In1kString-8    73.4µs ± 1%    56.5µs ± 1%  -22.92%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
In-8              552B ± 0%      632B ± 0%  +14.49%  (p=0.000 n=10+10)
In1k-8          29.1kB ± 0%    22.8kB ± 0%  -21.52%  (p=0.000 n=10+10)
In1kInt-8       37.1kB ± 0%    30.8kB ± 0%  -16.88%  (p=0.000 n=10+10)
In1kString-8    45.1kB ± 0%    38.8kB ± 0%  -13.88%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
In-8              11.0 ± 0%      12.0 ± 0%   +9.09%  (p=0.000 n=10+10)
In1k-8            11.0 ± 0%       6.0 ± 0%  -45.45%  (p=0.000 n=10+10)
In1kInt-8        1.01k ± 0%     1.01k ± 0%   -0.49%  (p=0.000 n=10+10)
In1kString-8     1.01k ± 0%     1.01k ± 0%   -0.49%  (p=0.000 n=10+10)
```